--- python/MemN2N.py
@@ -1,2 +1,0 @@
-import tensorflow as tf
-import numpy as np
